add castle as StartScreen. finished user story ab#530, completed task ab#531
