+++ Resources/Castle/Castle.txt
@@ -1,0 +1,51 @@
+▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
+▓..................................................................................................▓
+▓.....................███████. .█████...███▄.▄███▓.▄▄▄▄....██..█████...............................▓
+▓..........................▄▀..██▒..██...█▒▀█▀.██▒.█████▄..██..█...................................▓
+▓.........................▄▀....██░..██.▓██....▓██.▒██▒.▄██.██.▒███................................▓
+▓.......................▄▀......██...██..██....▒██.▒██░█▀...██.▒▓█.................................▓
+▓......................███████▒..████▓...██▒...░██.░▓█..▀█..██..▒████..............................▓
+▓....................................................▓███▀.........................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓......................▄████▄...▄▄▄........██████.▄▄▄█████▓.██.....▓█████..........................▓
+▓.....................▒██..▀█..▒████▄.....██.........██.....██.....██..............................▓
+▓....................▒▓█....▄.▒██...█▄.....██▄......██.....██░.....███.............................▓
+▓...................▒▓▓▄.▄██▒░██▄▄▄▄██.......██.....██....▒██░.....██..▄...........................▓
+▓...................▒.▓███▀ . ▓█...▓██..██████.....██....░██████▒..████▒...........................▓
+▓...................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓ ..........................................▒▒........▒▒...........................................▓
+▓...........................................▒.▒.......▒.▒..........................................▓
+▓...........................................▒..▒......▒..▒.........................................▓
+▓...........................................▒▒▒▒▒.....▒▒▒▒▒........................................▓
+▓...........................................▒.........▒............................................▓
+▓.................................▒.▒.▒.▒...▒.........▒........▒.▒.▒.▒.............................▓
+▓.................................▒▒▒▒▒▒▒...▒.........▒........▒▒▒▒▒▒▒.............................▓
+▓.................................▒▒▒▒▒▒▒.▒.▒.▒.▒.▒.▒.▒.▒.▒.▒..▒▒▒▒▒▒▒.............................▓
+▓.................................▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒.............................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒░░░░░░▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒░▒██▒░▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒░▒██▒░▒▒▒▒▒▒▒▒...................................▓
+▓......................................▒▒▒▒▒▒▒▒▒▒▒░▒██▒░▒▒▒▒▒▒▒▒...................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓..................................................................................................▓
+▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
