+++ src/main/java/com/zombiecastlerush/gui/AppMain.java
@@ -35,6 +35,7 @@ class AppMain extends JFrame implements KeyListener {
         AppMain app = new AppMain();
         app.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         app.setVisible(true);
+
     }
 
     private static String readWelcome(){
