+++ src/main/java/com/zombiecastlerush/gui/screens/LoseScreen.java
@@ -11,6 +11,6 @@ public class LoseScreen implements Screen {
     }
 
     public Screen respondToUserInput(KeyEvent key) {
-        return key.getKeyCode() == KeyEvent.VK_ENTER ? new PlayScreen() : this;
+        return key.getKeyCode() == KeyEvent.VK_ENTER ? new StartScreen() : this;
     }
 }
